+++ blinksy-desktop/examples/3d-arcs.rs
@@ -1,5 +1,5 @@
 use blinksy::{
-    layout::{Shape3d, Vec3},
+    layout::{Layout3d, Shape3d, Vec3},
     layout3d,
     patterns::noise::{noise_fns, Noise3d, NoiseParams},
     ControlBuilder,
@@ -63,7 +63,7 @@ layout3d!(
 fn main() {
     Desktop::new_3d::<TunnelLayout>().start(|driver| {
         let mut control = ControlBuilder::new_3d()
-            .with_layout::<TunnelLayout>()
+            .with_layout::<TunnelLayout, { TunnelLayout::PIXEL_COUNT }>()
             .with_pattern::<Noise3d<noise_fns::Perlin>>(NoiseParams {
                 ..Default::default()
             })