+++ blinksy-desktop/examples/3d-cube-volume-rainbow.rs
@@ -41,7 +41,7 @@ impl Layout3d for CubeVolumeLayout {
 fn main() {
     Desktop::new_3d::<CubeVolumeLayout>().start(|driver| {
         let mut control = ControlBuilder::new_3d()
-            .with_layout::<CubeVolumeLayout>()
+            .with_layout::<CubeVolumeLayout, { CubeVolumeLayout::PIXEL_COUNT }>()
             .with_pattern::<Rainbow>(RainbowParams::default())
             .with_driver(driver)
             .build();