+++ esp/gledopto/examples/ws2812-volume-cube.rs
@@ -46,7 +46,7 @@ fn main() -> ! {
     let p = board!();
 
     let mut control = ControlBuilder::new_3d()
-        .with_layout::<VolumeCubeLayout>()
+        .with_layout::<VolumeCubeLayout, { VolumeCubeLayout::PIXEL_COUNT }>()
         .with_pattern::<Noise3d<noise_fns::Perlin>>(NoiseParams {
             time_scalar: 0.25 / 1e3,
             position_scalar: 0.25,